A example of calibrating an all-sky image
This is an example used for the calibration of an all-sky image. It primarily utilizes HEALPix for the division and image transformation of the full-sky image so that Astrometry.net can identify stars, and then uses GWCS for image calibration of the registered stars. In addition to Astrometry.net and SExtractor, the Python packages used include numpy, astropy, healpy, scipy, gwcs, asdf, zoneinfo, datetime, os, and matplotlib.